pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "mcp>=1.13.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "sap-ai-sdk-gen[all]>=5.4.5",
]
